Bugfix choosing a partition number in S-INDD
The presented approach fails for attribute values
such as '2325029'. The hashcode of this string
and subsequently the partition number is then
negative, representing a non-valid array index.

The fix is to make the hashcode absolute. This
should retain all desired properties, e.g.,
that similar values are still assigned the same
partition.