Fix query semnatics with regards to NULL values
Entirely ignoring all tuples with at least one
NULL component may lead to false-positive INDs
since the appearance of a NULL value discards
the entire tuple from processing.

Instead all tuples with at least one component
which is not NULL should be processed. The
appearance of a NULL value in a record then
discards the IND early on since it is not
equal to any other value, including NULL.

A tuple entirely consisting of NULLs is
discarded from the left hand side since
it is the only case were we still can
safely conclude that the IND holds
without it (exactly when there is
another tuple with a non-NULL value
which is also part of RHS).